test: externalize turn executor white-box suite
